backend/requirements.txt
Flask>=2.3
flask-cors>=4.0
supabase>=2.4.0
PyJWT>=2.8.0
python-dotenv>=1.0.1
